// src/schemas/balancer-api.ts
export interface paths {
// ...
	"/gateways/{uuid}/config": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * Return YAML config for Traefik Gateway
		 * @description This endpoint is called from CloudGateway Automation only. So we are not going to provide ability to Account Admins to see a Gateway Config
		 *
		 */
		get: {
			parameters: {
				query?: never;
				header?: never;
				path: {
					/** @description Used for simplified authorization for some internal requests from CloudGateway to config.
					 *      */
					uuid: components["parameters"]["CloudGatewayId"];
				};
				cookie?: never;
			};
			requestBody?: never;
			responses: {
				/** @description Operation successful */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"text/yaml": components["schemas"]["GatewayConfig"];
					};
				};
				404: components["responses"]["not_found"];
			};
		};
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
	"/certificates": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * List of certificates attached to an account gateways
		 * @description This endpoint is called to retrieve the list of certificates attached to an account's gateways.
		 *     Role-based access:
		 *       - `operator` role can see all certificates in location
		 *       - `admin`/`owner` role can see all certificates in account
		 *       - `member` role can see only own certificates
		 *
		 */
		get: operations["certificates_list"];
		put?: never;
		/**
		 * Create a new certificate
		 * @description This endpoint is used for creating a new SSL certificate and uploading it to a **CloudGateway**. The server automatically checks the validity of the uploaded certificate. After the certificate has been successfully uploaded via the API, CloudGateway Automation automatically downloads and applies it to a route defined in the hostnames field.
		 *     Role-based access:
		 *       - `operator` role can create certificate in any account per location
		 *       - `admin`/`owner` role can create certificate only in the account space
		 *       - `member` role can create certificate only in the account space
		 *
		 */
		post: operations["certificates_create"];
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/services": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * Show all existing certificates
		 * @description This endpoint is called to retrieve the list of services along with information about VMs' IP addresses. It is typically used for fetching the service_id when users create routes using the API. The information is synchronized with the Compute application automatically every 15 minutes.
		 *     Role-based access:
		 *       - `operator` role can see all services in the location
		 *       - `admin`/`owner` role can see all services in account
		 *       - `member` role can see only own services
		 *
		 */
		get: operations["services_list"];
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
}
export type webhooks = Record<string, never>;
export interface components {
	schemas: {
		error: {
			/**
			 * @description HTTP code
			 * @example 500
			 */
			readonly status: number;
			data: {
				/**
				 * @description Error message description. Shown in UI as is, so make sure it understandable by ordinary user.
				 * @example Server error
				 */
				readonly message?: string;
			};
		};
		GatewayCollection: components["schemas"]["Gateway"][];
		Gateway: {
			/** @example 1234 */
			readonly id?: number;
			/**
			 * @description Name of gateway. No real technical restriction on names Validate: alphanumeric/dash/underscore
			 *
			 * @example vpn-default
			 */
			name: string;
// ...
		};
// ...
		Route: {
			route: {
				/** @example 1 */
				readonly id?: number;
				/** @example Route */
				name?: string;
				/** @example test.loc.icdc.io */
				hostname?: string;
// ...
				healthcheck_enabled?: boolean;
				cloud_gateway: {
					account: string;
					cloudgw_id: string;
					cloudgw_instance: string;
					created_at: string;
					id: number;
					name: string;
					updated_at: string;
				};
				healthcheck?: {
					/** @example / */
					path?: string;
					/** @example https */
					scheme?: string;
					/** @example check-cors.loc.icdc.io */
					hostname?: string;
					/** @example 8080 */
					port?: number;
					/** @example 15 */
					interval?: number;
					/** @example 3 */
					timeout?: number;
					/** @example {
					 *       "x-icdc-account": "devel",
					 *       "x-icdc-role": "admin"
					 *     } */
					headers?: Record<string, string>;
					/** @example POST */
					method?: string;
					/** @example true */
					follow_redirects?: boolean;
				};
			};
		};
		Route_POST: {
			route: Omit<components["schemas"]["Route"]["route"], "cloud_gateway">;
		};
	};
	responses: {
		/** @description This API gateway error means selected operation is not supported */
		apigw_not_allowed: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": string;
			};
		};
		/** @description Server error. */
		server_error: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Request is not authorized, usually because of: bad credentials, expired JWT token, account role is not assigned to authenticated user.
		 *      */
		unauthorized: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Request authorized, but current user doesn't have access rights to perform action.
		 *      */
		forbidden: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Requested object is not found
		 *      */
		not_found: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Unprocessable entity.
		 *      */
		unprocessable_entity: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description A list of ALB Gateways attached to account's infrastructure */
		alb_list: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["GatewayCollection"];
			};
		};
	};
	parameters: {
		/** @description Name of the account (5-character string) */
		Account: string;
		/** @description Role name */
		Role: string;
		/** @description Used for simplified authorization for some internal requests from CloudGateway to config.
		 *      */
		CloudGatewayId: string;
	};
	requestBodies: never;
	headers: never;
	pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
	alb_list: {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			200: components["responses"]["alb_list"];
			401: components["responses"]["unauthorized"];
			500: components["responses"]["server_error"];
		};
	};
	alb_attach: {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody: {
			content: {
				"application/json": {
					/** @example alb-default */
					name?: string;
					/** @example a843bf2a-7353-563e-a220-f59949849dba */
					cloudgw_id?: string;
					/** @example acc */
					account?: string;
				};
			};
		};
		responses: {
			200: components["responses"]["alb_list"];
			401: components["responses"]["unauthorized"];
			405: components["responses"]["apigw_not_allowed"];
			422: components["responses"]["unprocessable_entity"];
		};
	};
// ...
}